requirements.txt
streamlit==1.53.0
tensorflow==2.20.0
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==12.1.0
huggingface_hub
